Add the Iodine server + Websocket support feature
I'm the author of [iodine](https://github.com/boazsegev/iodine) and I love using it in my apps, especially with it's native Websocket support.

I thought I'd share the joy.

I also added a suggestion for the Static File server feature. Both iodine and passenger support this (iodine supports pre-compressed `gzip` detection as well). It's something people sometimes look for, so I think it would be helpful to mention.

While I was editing the file, I also added the Websocket support column.

It might be a bit early to add it, but iodine supports native Websockets and I was [discussing it a while back](https://github.com/boazsegev/iodine/issues/6) both with the Puma and the Passenger teams and we came up with an [outline for the Websocket Rack Specification](https://github.com/boazsegev/iodine/blob/master/SPEC-Websocket-Draft.md)...

I guess each server will probably implement variations on the theme proposed, but it might take a bit of time before some [road blocks related to the C10K issue](https://github.com/puma/puma/pull/1056) are cleared.

Anyway, I was hoping that adding the feature to the list will push things forward.

Thanks for a great listing and all the details.

B.